Moved lodin and upload to cloud dir
