+++ mscore/cloud/cloud.cmake
@@ -1,0 +1,16 @@
+
+set(CLOUD_SRC
+    ${CMAKE_CURRENT_LIST_DIR}/logindialog.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/logindialog.h
+    ${CMAKE_CURRENT_LIST_DIR}/loginmanager.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/loginmanager.h
+    ${CMAKE_CURRENT_LIST_DIR}/loginmanager_p.h
+    ${CMAKE_CURRENT_LIST_DIR}/uploadscoredialog.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/uploadscoredialog.h
+    )
+
+
+set(CLOUD_UI
+    ${CMAKE_CURRENT_LIST_DIR}/logindialog.ui
+    ${CMAKE_CURRENT_LIST_DIR}/uploadscoredialog.ui
+    )
