+++ mscore/cloud/logindialog.h
