+++ mscore/cloud/logindialog.ui
