+++ mscore/cloud/loginmanager.cpp
