+++ mscore/cloud/loginmanager.h
