+++ mscore/cloud/loginmanager_p.h
